+++ next.config.mjs
@@ -2,6 +2,20 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     reactStrictMode: true,
+    images: {
+        remotePatterns: [
+            {
+                protocol: 'https',
+                hostname: 'lh3.googleusercontent.com',
+                pathname: '/**',
+            },
+            {
+                protocol: 'https',
+                hostname: 'googleusercontent.com',
+                pathname: '/**',
+            },
+        ],
+    },
 };
 
 export default nextConfig;